Перенос с jdbc на jpa.

+++ spring-06/src/main/java/my/spring/service/book/BookService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface BookService {
 
-    int getBooksCount();
+    long getBooksCount();
 
     Book getBookById(long id);
